feat: reafctoring and add descriptions
