+++ .gitignore
@@ -1,4 +1,4 @@
-data/*
+add_session_id/data/*
 
 # DotEnv configuration
 .env
