+++ README.md
@@ -1,5 +1,5 @@
-# add_session_id_project
-test task for the interview
+# Calculate and add "session_id" to dataframe
+Пример работы с pandas
 
 ![image](https://user-images.githubusercontent.com/40138357/199262774-42d5806c-b802-4a4e-8643-f572bbaa3b6d.png)
