+++ add_session_id/config/config.py
@@ -2,10 +2,8 @@ import os
 import sys
 from pathlib import Path
 from typing import Optional
-from datetime import datetime
 from dotenv import dotenv_values
 from loguru import logger as base_logger
-from functools import wraps
 
 
 def get_env_dict(env_dir: Optional[Path]):
@@ -20,7 +18,7 @@ def get_logger(logging_level: str, log_dir: Optional[Path] = None, logging_dir_l
     base_logger.remove()
     base_logger.add(sys.stderr, level=logging_level)
 
-    # если указать путь, то логи будут записываться не только в std.err
+    # если указать путь, то логи будут записываться не только в stderr
     if log_dir:
         base_logger.add(log_dir / "{time}.log", level=logging_dir_level)
 
@@ -33,22 +31,3 @@ logger = get_logger(
     log_dir=env_dict.get("LOGGING_DIR_LEVEL", None),
     logging_dir_level=env_dict.get("LOGGING_DIR_LEVEL", "DEBUG"),
 )
-
-
-def logger_dec(func_info: str) -> callable:
-    def decorator(func: callable) -> callable:
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-
-            start_time = datetime.now()
-            logger.info(f"start {func_info}")
-
-            res = func(*args, **kwargs)
-
-            logger.info(f"finish {func_info} ({datetime.now() - start_time})")
-
-            return res
-
-        return wrapper
-
-    return decorator
