+++ add_session_id/config/shared.env
@@ -5,7 +5,7 @@ N_ROWS=100_000_000
 START_TIME="2022-01-01 00:00:00" #"YYYY-MM-dd HH:mm:ss"
 END_TIME="2023-01-01 00:00:00"
 
-INPUT_FILE_PATH="./data/data.csv"
-OUTPUT_FILE_PATH="./data/new_data.csv"
+INPUT_FILE_PATH="./add_session_id/data/data.csv"
+OUTPUT_FILE_PATH="./add_session_id/data/new_data.csv"
 
 LOGGING_LEVEL="INFO"
