+++ run.sh
@@ -1,0 +1,3 @@
+pip install -e .
+pip install -r add_session_id/requirements/requirements.txt
+python add_session_id/main.py -sgen=False -sres=False -ff=False
