+++ setup.py
@@ -1,0 +1,9 @@
+from setuptools import find_packages, setup
+
+setup(
+    name="add_session_id",
+    packages=find_packages(),
+    version="0.0.1",
+    description="An example of Godun Roman's code",
+    license="MIT",
+)
